refactor completed - main responsibilities separated - no consideration given to maintenance nor coupling between classes as system is very small
