+++ src/main/java/dbutils/EntityManagerFactoryGenerator.java
@@ -1,0 +1,14 @@
+package dbutils;
+
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.Persistence;
+
+public class EntityManagerFactoryGenerator {
+    private static EntityManagerFactory emf;
+
+    public static EntityManagerFactory getFactory() {
+        if (emf == null) emf = Persistence.createEntityManagerFactory("default");
+        return emf;
+    }
+
+}
